TASK-2 (Project structure) - refactor: change user and article services to models 		modified:   autoload.php         modified:   src/App/Controllers/ArticleController.php         modified:   src/App/Controllers/AuthController.php         modified:   src/App/Controllers/HomeController.php         new file:   src/App/Models/Articles/ArticleModel.php         new file:   src/App/Models/Users/UserModel.php         deleted:    src/App/Services/ArticleService.php         deleted:    src/App/Services/UserService.php         modified:   src/App/container-definitions.php         modified:   src/App/functions.php         modified:   src/App/views/articles/create.php         modified:   src/App/views/articles/edit.php         modified:   src/App/views/index.php         new file:   src/Framework/Model.php         modified:   src/Framework/TemplateEngine.php
