+++ src/App/Controllers/AuthController.php
@@ -5,14 +5,16 @@ declare(strict_types=1);
 namespace App\Controllers;
 
 use Framework\TemplateEngine;
-use App\Services\{ValidatorService, UserService};
+use App\Services\ValidatorService;
+use App\Models\Users\UserModel;
+
 
 class AuthController
 {
     public function __construct(
         private TemplateEngine $view,
         private ValidatorService $validatorService,
-        private UserService $userService
+        private UserModel $userModel
     ) {
     }
 
@@ -25,9 +27,9 @@ class AuthController
     {
         $this->validatorService->validateRegister($_POST);
 
-        $this->userService->isEmailTaken($_POST['email']);
+        $this->userModel->isEmailTaken($_POST['email']);
 
-        $this->userService->create($_POST);
+        $this->userModel->create($_POST);
         redirectTo('/');
     }
 
@@ -40,14 +42,14 @@ class AuthController
     {
         $this->validatorService->validateLogin($_POST);
 
-        $this->userService->login($_POST);
+        $this->userModel->login($_POST);
 
         redirectTo('/');
     }
 
     public function logout()
     {
-        $this->userService->logout();
+        $this->userModel->logout();
 
         redirectTo('/login');
     }
