+++ src/Framework/Model.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Framework;
+
+class Model
+{
+  public function __construct(protected Database $db)
+  {
+  }
+}
